pyproject.toml
[project]
name = "aicmdb"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "alinx", email = "itadminlx@163.com"},
]
dependencies = [
    "Django==5.0.6",
    "djangorestframework==3.15.1",
    "mysqlclient==2.2.4",
    "crypto",
    "django-cors-headers",
    "django-shortuuidfield",
    "PyJWT",
    "aicmdb @ file:///${PROJECT_ROOT}/",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
